mad improved

+++ kernels.c
@@ -390,9 +390,9 @@ void convolution(int dim, pixel *src, pixel *ker, unsigned *dst)
     unsigned limit = dim - 7;
     pixel *kerRow = ker;
     unsigned val = 0;
-    unsigned short k0r, k1r, k2r, k3r, k4r, k5r, k6r, k7r;
-    unsigned short k0g, k1g, k2g, k3g, k4g, k5g, k6g, k7g;
-    unsigned short k0b, k1b, k2b, k3b, k4b, k5b, k6b, k7b;
+    unsigned k0r, k1r, k2r, k3r, k4r, k5r, k6r, k7r;
+    unsigned k0g, k1g, k2g, k3g, k4g, k5g, k6g, k7g;
+    unsigned k0b, k1b, k2b, k3b, k4b, k5b, k6b, k7b;
     for (k = 0; k < 8; k += 1, kerRow += 8)
     {
 
@@ -455,7 +455,6 @@ void register_conv_functions()
 {
     add_conv_function(&naive_conv, naive_conv_descr);
     add_conv_function(&convolution, convolution_descr);
-    // add_conv_function(&convolutionMad, convolutionMad_descr);
     add_conv_function(&convolutionMustafaV2, convolutionMustafaV2_descr);
     add_conv_function(&convolutionMustafaVf, convolutionMustafaVf_descr);
